ENVIA VIDEO COM SUCESSO,ainda perde seq as vezes

+++ Packet.h
@@ -110,7 +110,7 @@ char *get_type_packet(unsigned char* buffer);
 /* ================ funcoes auxiliares ================ */
 void read_packet(unsigned char *buffer);
 int is_our_packet(unsigned char *buffer);
-int is_our_mask(unsigned int *buffer);
+int is_our_mask(unsigned long *buffer);
 int is_valid_type(int tipo);
 void print_bytes(char *nome,unsigned char *buf, int n);
